delete and update functions done

+++ app/src/main/java/com/example/passwordmanager/screens/login/LoginFragment.kt
@@ -33,7 +33,7 @@ class LoginFragment : NavigationFragment<FragmentLoginBinding>(R.layout.fragment
         }
 
         viewBinding.btnConfirm.setOnClickListener {
-            viewModel.login(viewBinding.loginFragmentEtPassword.text.toString().toLong())
+            viewModel.login(viewBinding.loginFragmentEtPassword.text.toString())
         }
